+++ index.js
@@ -1,0 +1,3 @@
+// En este archivo se exporta la librería
+const cardify = require("src/cardify.js");
+module.exports = cardify;
